Caching of Ropeforce calculation
The biggest amount of calculations of a rope's force will be cached now,
so it won't be calculated twice (since there are two objects which need
a force vector in every simulation step).

When the program ran for a minute, Vector3::normalize() took 1.5% of
calculation time and sin() took 1.4%.
Not the percentages shrunk to 1% and 0.8%.
This could come in handy for bigger amounts of ropes. :-)